fix: Storage now displaying correctly
(had to add context to formatting functions)

+++ app/src/main/java/catmonit/app/ui/storage/StorageState.java
@@ -1,19 +1,19 @@
 package catmonit.app.ui.storage;
 
 public class StorageState {
-    private final int total_space_bytes;
-    private final int used_space_bytes;
+    private final long total_space_bytes;
+    private final long used_space_bytes;
 
-    public StorageState(int total_space_bytes, int used_space_bytes) {
+    public StorageState(long total_space_bytes, long used_space_bytes) {
         this.total_space_bytes = total_space_bytes;
         this.used_space_bytes = used_space_bytes;
     }
 
-    public int getTotal_space_bytes() {
+    public long getTotal_space_bytes() {
         return total_space_bytes;
     }
 
-    public int getUsed_space_bytes() {
+    public long getUsed_space_bytes() {
         return used_space_bytes;
     }
